renamed and moved some files for a better architecture
