+++ app/src/main/java/eu/kudan/ar/di/ProjectInitializer.java
@@ -16,12 +16,12 @@ import commonlib.location.rotation.Rotator;
 import commonlib.location.rotation.VectorRotator;
 import commonlib.storage.FloatMeanRingBuffer;
 import commonlib.storage.LocationMeanRingbufferImp;
-import commonlib.storage.RingBufferImpl;
+import commonlib.storage.MeanRingBufferAbstract;
 import eu.kudan.ar.ar.ARBuildingsPositioner;
-import eu.kudan.ar.ar.position.GPSBuildingPositioner;
-import eu.kudan.ar.ar.position.location.LocationFilter;
-import eu.kudan.ar.ar.position.location.NorthSensorListener;
-import eu.kudan.ar.ar.position.location.PhysicalNorthInitializer;
+import commonlib.GPSBuildingPositioner;
+import commonlib.location.LocationFilter;
+import commonlib.location.NorthSensorListener;
+import commonlib.location.PhysicalNorthInitializer;
 import eu.kudan.kudan.ARAPIKey;
 import eu.kudan.kudan.ARGyroManager;
 import eu.kudan.kudan.ARModelImporter;
@@ -50,9 +50,9 @@ public class ProjectInitializer {
         Texturizer blackTexture = new TexturizerModelBlack();
         BuildingModelSampleImpl buildingModel = new BuildingModelSampleImpl(model, blackTexture);
         LocationDistanceCalculator locationDistanceCalculator = new LocationDistanceCalculator();
-        RingBufferImpl<Location> locationMean = new LocationMeanRingbufferImp(10);
+        MeanRingBufferAbstract<Location> locationMean = new LocationMeanRingbufferImp(10);
         LocationFilter locationFilter = new LocationFilter(locationMean, locationManager);
-        RingBufferImpl<Float> angleRingBuffer = new FloatMeanRingBuffer(100);
+        MeanRingBufferAbstract<Float> angleRingBuffer = new FloatMeanRingBuffer(100);
         NorthSensorListener northSensorListener = new NorthSensorListener(sensorManager, angleRingBuffer, 100);
         Rotator<Vector3f> vector3fRotator = new VectorRotator(new Vector3f());
