+++ commonlib/src/main/java/commonlib/location/LocationFilter.java
@@ -1,4 +1,4 @@
-package eu.kudan.ar.ar.position.location;
+package commonlib.location;
 
 import android.annotation.SuppressLint;
 import android.location.Location;
